Support OpenSSL on Win32 and Arm windows platforms (#1061)
* Support OpenSSL on Win32 and Arm windows platforms

* A bit of imported target adjustments

* Add openssl x86 and arm64 builds

* Fix target properties

+++ src/platform/tls_openssl.c
@@ -341,7 +341,7 @@ QuicTlsAddHandshakeDataCallback(
         OpenSslAddHandshakeData,
         TlsContext->Connection,
         "Sending %llu handshake bytes (Level = %u)",
-        Length,
+        (uint64_t)Length,
         Level);
 
     if (Length + TlsState->BufferLength > 0xF000) {
